third_party/WebKit/LayoutTests/fast/dom/Window/replaceable-expected.txt

Tests that Replaceable attributes are writable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Object.getOwnPropertyDescriptor(window, "innerHeight").set is an instance of Function
PASS window.innerHeight is 42
PASS window.event is "foo"
PASS successfullyParsed is true

TEST COMPLETE
